Refactor: QuizDataCenterMediator 통합
- QuizCategory 별 QuizDataCenterMediator 를 통합
